+++ src/main/java/com/webank/webase/sign/aspect/LogAspect.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2014-2019  the original author or authors.
+ * Copyright 2014-2020  the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -15,6 +15,8 @@ package com.webank.webase.sign.aspect;
 
 import java.time.Duration;
 import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
@@ -22,9 +24,10 @@ import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Pointcut;
 import org.aspectj.lang.reflect.MethodSignature;
 import org.slf4j.Logger;
+import com.webank.webase.sign.util.JsonUtils;
 import org.springframework.stereotype.Component;
-import com.alibaba.fastjson.JSON;
 import com.webank.webase.sign.manager.LoggerManager;
+import org.springframework.validation.BindingResult;
 
 @Aspect
 @Component
@@ -45,8 +48,10 @@ public class LogAspect {
         String methodName = methodSignature.getName();
         Object[] args = point.getArgs();
         Logger logger = LoggerManager.getLogger(targetClass);
+        // log args of param in controller
+        // if args contains BindingResult(recursive of request entity and itself), stack over flow
         logger.debug("startTime:{} methodName:{} args:{}", startTime, methodName,
-            JSON.toJSONString(args));
+            JsonUtils.toJSONString(this.excludeBindingResult(args)));
         Object result = null;
         try {
             result = point.proceed();
@@ -55,10 +60,19 @@ public class LogAspect {
             throw throwable;
         }
 
-        String resultStr = Optional.ofNullable(result).map(r -> JSON.toJSONString(r)).orElse(null);
-        logger.debug("methodName:{} userTime:{} result:{}", methodName,
+        String resultStr = Optional.ofNullable(result).map(JsonUtils::toJSONString).orElse(null);
+        logger.debug("methodName:{} usedTime:{} result:{}", methodName,
             Duration.between(startTime, Instant.now()), resultStr);
         return result;
     }
 
+    private List<Object> excludeBindingResult(Object[] params) {
+        List<Object> retainParams = new ArrayList<>();
+        for (int index = 0; index < params.length; index++) {
+            if (!(params[index] instanceof BindingResult)) {
+                retainParams.add(params[index]);
+            }
+        }
+        return retainParams;
+    }
 }